Add ability to tail multiple readers
Updates the main loop to be able to read from multiple readers passed as
arguments. Moves the main.go entrypoint to the `cmd/` directory to allow
adding other go files to the root directory in the future. Also adds a
Makefile and a .gitignore to make it easier to build the project.